+++ nest/rollback.go
@@ -404,7 +404,7 @@ func (tx *RollbackTx) prepareCommitRecord() (CommitRecord, error) {
 		mutations[i] = canonical
 	}
 	record := CommitRecord{
-		ID: tx.id, Handler: tx.handler, RequestID: requestID, CreatedAt: time.Now().UnixNano(), Durability: uint8(tx.durability),
+		ID: tx.id, Handler: tx.handler, RequestID: requestID, CreatedAt: time.Now().UnixNano(), Durability: tx.durability,
 		Mutations: mutations,
 		Effects:   append([]Effect(nil), tx.effects...),
 		Receipts:  append([]dataengine.Receipt(nil), tx.receipts...),
